Remove vendor folder
The vendor folder had been initially added to allow interoperation
with older versions of Go and to make sure we had a copy of all deps
living under the same file tree.

Depending on the module users Go version, the first part may not
hold. As for making sure we keep a copy of the dependencies, it
should not be a problem if dependency repositories are not
completely removed.

Neither of these seem to be/should be an issue, so based on comment
https://github.com/marshallbrekka/go-u2fhost/pull/10#issuecomment-757624060
this commit removes vendor/