Support linux usb hid discovery (#10)
Optional support for discovering USB HID devices on Linux

Requires passing "-tags=hidraw" when building to enable the fix.

This commit changes the hid library used for accessing u2f USB devices
from karalabe/hid to bearsh/hid. bearsh/hid has updated karalabe's
libusb and hidapi code base with upstream libusb/libusb and
libusb/hidapi, which among other things incorporate the changes in
libusb/hidapi to support Usage Page and Usage on Linux with hidraw.
(see https://github.com/libusb/hidapi/pull/139/files) that fix
https://github.com/marshallbrekka/go-u2fhost/issues/1 when built with
"-tags=hidraw", while maintaining the old behaviour when built without
it. The fork includes fixes and enhancements related to vendoring
modules.

The commit also fixes the module directive in go.mod and adds the
vendor directory resulting from executing "go mod vendor"

* Remove vendor folder

The vendor folder had been initially added to allow interoperation
with older versions of Go and to make sure we had a copy of all deps
living under the same file tree.

Depending on the module users Go version, the first part may not
hold. As for making sure we keep a copy of the dependencies, it
should not be a problem if dependency repositories are not
completely removed.

Neither of these seem to be/should be an issue, so based on comment
https://github.com/marshallbrekka/go-u2fhost/pull/10#issuecomment-757624060
this commit removes vendor/

Co-authored-by: Brendan Roy <br3ndanr@gmail.com>